some error handling

+++ app/src/services/userBlockService.ts
@@ -1,3 +1,4 @@
+import { check } from 'drizzle-orm/gel-core';
 import type {
 	PlacementQuizDto,
 	CreatePlacementQuizDto,
@@ -6,17 +7,20 @@ import type {
 	UpdateUserBlockDto,
 	UserBlockDto
 } from '../db/schema';
+import { NotFoundError, UnauthorizedError } from '../errors/AppError';
 
 import { PlacementQuizRepository } from '../repositories/placementQuizRepository';
 import { UserBlockRepository } from '../repositories/userBlockRepository';
+import { UserRepository } from '../repositories/userRepository';
 import type { Transaction } from '../types';
-import { NotFoundError } from './utils/notFoundError';
 
 export class UserBlockService {
 	private repo: UserBlockRepository;
+	private userRepo: UserRepository;
 
 	constructor() {
 		this.repo = new UserBlockRepository();
+		this.userRepo = new UserRepository();
 	}
 
 	async getById(id: string, tx?: Transaction): Promise<UserBlockDto> {
@@ -29,13 +33,18 @@ export class UserBlockService {
 		return await this.repo.create(data, tx);
 	}
 
-	async update(id: string, data: UpdateUserBlockDto, tx?: Transaction): Promise<UserBlockDto> {
+	async update(
+		id: string,
+		data: UpdateUserBlockDto,
+		tx?: Transaction
+	): Promise<UserBlockDto> {
 		const item = await this.repo.update(id, data, tx);
 		if (!item) throw new NotFoundError(`UserBlock with id ${id} not found`);
 		return item;
 	}
 
-	async delete(id: string, tx?: Transaction): Promise<UserBlockDto> {
+	async delete(id: string, userEmail: string, tx?: Transaction): Promise<UserBlockDto> {
+		await this.checkUserIsOwnerOfUserBlock(id, userEmail, tx);
 		const item = await this.repo.delete(id, tx);
 		if (!item) throw new NotFoundError(`UserBlock with id ${id} not found`);
 		return item;
@@ -50,4 +59,17 @@ export class UserBlockService {
 		return item;
 	}
 
+	async checkUserIsOwnerOfUserBlock(
+		userBlockId: string,
+		userEmail: string,
+		tx?: Transaction
+	): Promise<void> {
+		const user = await this.userRepo.getByEmail(userEmail, tx);
+		if (!user) throw new NotFoundError(`User with email ${userEmail} not found`);
+		const userBlock = await this.repo.getById(userBlockId, tx);
+		if (!userBlock) throw new NotFoundError(`UserBlock with id ${userBlockId} not found`);
+
+		if (user.id !== userBlock.userId)
+			throw new UnauthorizedError(`User is not the owner of UserBlock ${userBlockId}`);
+	}
 }
